Add new api response handler class and change tests
