+++ src/main/java/uk/gov/companieshouse/disqualifiedofficers/search/handler/ApiResponseHandler.java
@@ -1,0 +1,37 @@
+package uk.gov.companieshouse.disqualifiedofficers.search.handler;
+
+import java.util.Map;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
+import uk.gov.companieshouse.disqualifiedofficers.search.exception.NonRetryableErrorException;
+import uk.gov.companieshouse.disqualifiedofficers.search.exception.RetryableErrorException;
+import uk.gov.companieshouse.logging.Logger;
+
+public class ApiResponseHandler {
+
+    /**
+     * Handle responses by logging result and throwing any exceptions.
+     */
+    public void handleResponse(
+            final ResponseStatusException ex,
+            final HttpStatus httpStatus,
+            final String logContext,
+            final String msg,
+            final Map<String, Object> logMap,
+            Logger logger)
+            throws NonRetryableErrorException, RetryableErrorException {
+        logMap.put("status", httpStatus.toString());
+        if (HttpStatus.BAD_REQUEST == httpStatus) {
+            // 400 BAD REQUEST status cannot be retried
+            logger.errorContext(logContext, msg, null, logMap);
+            throw new NonRetryableErrorException(msg);
+        } else if (httpStatus.is4xxClientError() || httpStatus.is5xxServerError()) {
+            // any other client or server status can be retried
+            logger.errorContext(logContext, msg + ", retry", null, logMap);
+            throw new RetryableErrorException(msg);
+        } else {
+            logger.debugContext(logContext, msg, logMap);
+        }
+    }
+    
+}
