+++ src/main/java/uk/gov/companieshouse/disqualifiedofficers/search/processor/ResourceChangedProcessor.java
@@ -5,7 +5,6 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.messaging.Message;
@@ -13,8 +12,8 @@ import org.springframework.stereotype.Component;
 
 import uk.gov.companieshouse.api.disqualification.OfficerDisqualification;
 import uk.gov.companieshouse.api.model.ApiResponse;
-import uk.gov.companieshouse.disqualifiedofficers.search.exception.NonRetryableErrorException;
 import uk.gov.companieshouse.disqualifiedofficers.search.exception.RetryableErrorException;
+import uk.gov.companieshouse.disqualifiedofficers.search.handler.ApiResponseHandler;
 import uk.gov.companieshouse.disqualifiedofficers.search.service.api.ApiClientService;
 import uk.gov.companieshouse.disqualifiedofficers.search.transformer.ElasticSearchTransformer;
 import uk.gov.companieshouse.logging.Logger;
@@ -42,7 +41,6 @@ public class ResourceChangedProcessor {
     }
 
     public void processResourceChanged(Message<ResourceChangedData> message) {
-        try {
             final Map<String, Object> logMap = new HashMap<>();
             ResourceChangedData payload = message.getPayload();
             final String logContext = payload.getContextId();
@@ -54,13 +52,16 @@ public class ResourceChangedProcessor {
             String officerId = Stream.of( elasticSearchData.getLinks().getSelf().split("/") )
                     .reduce( (first,last) -> last ).get();
 
+        try {
             final ApiResponse<Void> response =
                     apiService.putDisqualificationSearch(logContext, officerId, elasticSearchData);
             logger.infoContext(
                     logContext,
                     String.format("Process disqualification for officer with id [%s]", officerId),
                     null);
-            handleResponse(null, HttpStatus.valueOf(response.getStatusCode()),
+
+            ApiResponseHandler apiResponseHandler = new ApiResponseHandler();
+            apiResponseHandler.handleResponse(null, HttpStatus.valueOf(response.getStatusCode()),
                     logContext,"Response received from search api", logMap, logger);
         } catch (RetryableErrorException ex) {
             retryResourceChangedMessage(message);
@@ -74,31 +75,6 @@ public class ResourceChangedProcessor {
     }
 
     private void handleErrorMessage(Message<ResourceChangedData> message) {
-    }
-
-    /**
-     * Handle responses by logging result and throwing any exceptions.
-     */
-    public void handleResponse(
-            final ResponseStatusException ex,
-            final HttpStatus httpStatus,
-            final String logContext,
-            final String msg,
-            final Map<String, Object> logMap,
-            Logger logger)
-            throws NonRetryableErrorException, RetryableErrorException {
-        logMap.put("status", httpStatus.toString());
-        if (HttpStatus.BAD_REQUEST == httpStatus) {
-            // 400 BAD REQUEST status cannot be retried
-            logger.errorContext(logContext, msg, null, logMap);
-            throw new NonRetryableErrorException(msg);
-        } else if (httpStatus.is4xxClientError() || httpStatus.is5xxServerError()) {
-            // any other client or server status can be retried
-            logger.errorContext(logContext, msg + ", retry", null, logMap);
-            throw new RetryableErrorException(msg);
-        } else {
-            logger.debugContext(logContext, msg, logMap);
-        }
     }
 
 }
